# aspnetcore/src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.2",
  "info" : {
    "title" : "Dilmah-Admin-BFF",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "/"
  } ],
  "paths" : {
...
    "/orders" : {
      "get" : {
        "tags" : [ "Orders" ],
        "summary" : "Get Orders",
        "parameters" : [ {
          "name" : "Auth-token",
          "in" : "header",
          "description" : "Authorization bearea token",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "success",
            "content" : {
              "application/json" : {
                "examples" : {
                  "success" : {
                    "value" : "\n    {\n        \"id\": 21,\n        \"checkoutCompletedAt\": \"2017-02-15T13:31:33+0100\",\n        \"number\": \"000000021\",\n        \"items\": [\n            {\n                \"id\": 74,\n                \"quantity\": 1,\n                \"unitPrice\": 100000,\n                \"total\": 100000,\n                \"units\": [\n                    {\n                        \"id\": 228,\n                        \"adjustments\": [],\n                        \"adjustmentsTotal\": 0,\n                        \"_links\": {\n                            \"order\": {\n                                \"href\": \"/api/v1/orders/21\"\n                            }\n                        }\n                    }\n                ],\n                \"unitsTotal\": 100000,\n                \"adjustments\": [],\n                \"adjustmentsTotal\": 0,\n                \"variant\": {\n                    \"id\": 331,\n                    \"code\": \"MEDIUM_MUG_CUP\",\n                    \"optionValues\": [\n                        {\n                            \"name\": \"Mug type\",\n                            \"code\": \"mug_type_medium\"\n                        }\n                    ],\n                    \"position\": 2,\n                    \"translations\": {\n                        \"en_US\": {\n                            \"locale\": \"en_US\",\n                            \"id\": 331,\n                            \"name\": \"Medium Mug\"\n                        }\n                    },\n                    \"version\": 1,\n                    \"onHold\": 0,\n                    \"onHand\": 10,\n                    \"tracked\": false,\n                    \"channelPricings\": {\n                        \"US_WEB\": {\n                            \"channelCode\": \"US_WEB\",\n                            \"price\": 100000\n                        }\n                    },\n                    \"_links\": {\n                        \"self\": {\n                            \"href\": \"/api/v1/products/5/variants/331\"\n                        },\n                        \"product\": {\n                            \"href\": \"/api/v1/products/5\"\n                        }\n                    }\n                },\n                \"_links\": {\n                    \"product\": {\n                        \"href\": \"/api/v1/products/5\"\n                    },\n                    \"variant\": {\n                        \"href\": \"/api/v1/products/5/variants/331\"\n                    }\n                }\n            }\n        ],\n        \"itemsTotal\": 100000,\n        \"adjustments\": [\n            {\n                \"id\": 252,\n                \"type\": \"shipping\",\n                \"label\": \"DHL Express\",\n                \"amount\": 3549\n            }\n        ],\n        \"adjustmentsTotal\": 3549,\n        \"total\": 103549,\n        \"state\": \"new\",\n        \"customer\": {\n            \"id\": 1,\n            \"email\": \"shop@example.com\",\n            \"emailCanonical\": \"shop@example.com\",\n            \"firstName\": \"John\",\n            \"lastName\": \"Doe\",\n            \"gender\": \"u\",\n            \"user\": {\n                \"id\": 1,\n                \"username\": \"shop@example.com\",\n                \"usernameCanonical\": \"shop@example.com\",\n                \"roles\": [\n                    \"ROLE_USER\"\n                ],\n                \"enabled\": true\n            },\n            \"_links\": {\n                \"self\": {\n                    \"href\": \"/api/v1/customers/1\"\n                }\n            }\n        },\n        \"channel\": {\n            \"id\": 1,\n            \"code\": \"US_WEB\",\n            \"name\": \"US Web Store\",\n            \"hostname\": \"localhost\",\n            \"color\": \"MediumPurple\",\n            \"createdAt\": \"2017-02-14T11:10:02+0100\",\n            \"updatedAt\": \"2017-02-14T11:10:02+0100\",\n            \"enabled\": true,\n            \"taxCalculationStrategy\": \"order_items_based\",\n            \"_links\": {\n                \"self\": {\n                    \"href\": \"/api/v1/channels/1\"\n                }\n            }\n        },\n        \"shippingAddress\": {\n            \"id\": 71,\n            \"firstName\": \"Frederick D.\",\n            \"lastName\": \"Gregory\",\n            \"countryCode\": \"US\",\n            \"street\": \"300 E St SW\",\n            \"city\": \"’Washington\",\n            \"postcode\": \"DC 20546\",\n            \"createdAt\": \"2017-02-14T11:55:40+0100\",\n            \"updatedAt\": \"2017-02-14T17:00:17+0100\"\n        },\n        \"billingAddress\": {\n            \"id\": 72,\n            \"firstName\": \"Frederick D.\",\n            \"lastName\": \"Gregory\",\n            \"countryCode\": \"US\",\n            \"street\": \"300 E St SW\",\n            \"city\": \"’Washington\",\n            \"postcode\": \"DC 20546\",\n            \"createdAt\": \"2017-02-14T11:55:40+0100\",\n            \"updatedAt\": \"2017-02-14T17:00:17+0100\"\n        },\n        \"payments\": [\n            {\n                \"id\": 21,\n                \"method\": {\n                    \"id\": 2,\n                    \"code\": \"bank_transfer\",\n                    \"createdAt\": \"2017-02-14T11:10:02+0100\",\n                    \"updatedAt\": \"2017-02-14T11:10:02+0100\",\n                    \"channels\": [\n                        {\n                            \"id\": 1,\n                            \"code\": \"US_WEB\",\n                            \"name\": \"US Web Store\",\n                            \"hostname\": \"localhost\",\n                            \"color\": \"MediumPurple\",\n                            \"createdAt\": \"2017-02-14T11:10:02+0100\",\n                            \"updatedAt\": \"2017-02-14T11:10:02+0100\",\n                            \"enabled\": true,\n                            \"taxCalculationStrategy\": \"order_items_based\",\n                            \"_links\": {\n                                \"self\": {\n                                    \"href\": \"/api/v1/channels/1\"\n                                }\n                            }\n                        }\n                    ],\n                    \"_links\": {\n                        \"self\": {\n                            \"href\": \"/api/v1/payment-methods/bank_transfer\"\n                        }\n                    }\n                },\n                \"amount\": 103549,\n                \"state\": \"new\",\n                \"createdAt\": \"2017-02-14T11:53:41+0100\",\n                \"updatedAt\": \"2017-02-15T13:31:33+0100\",\n                \"_links\": {\n                    \"self\": {\n                        \"href\": \"/api/v1/payments/21\"\n                    },\n                    \"payment-method\": {\n                        \"href\": \"/api/v1/payment-methods/bank_transfer\"\n                    },\n                    \"order\": {\n                        \"href\": \"/api/v1/orders/21\"\n                    }\n                }\n            }\n        ],\n        \"shipments\": [\n            {\n                \"id\": 21,\n                \"state\": \"ready\",\n                \"method\": {\n                    \"id\": 2,\n                    \"code\": \"dhl_express\",\n                    \"category_requirement\": 1,\n                    \"calculator\": \"flat_rate\",\n                    \"configuration\": {\n                        \"US_WEB\": {\n                            \"amount\": 3549\n                        }\n                    },\n                    \"createdAt\": \"2017-02-14T11:10:02+0100\",\n                    \"updatedAt\": \"2017-02-14T11:10:02+0100\",\n                    \"enabled\": true,\n                    \"_links\": {\n                        \"self\": {\n                            \"href\": \"/api/v1/shipping-methods/dhl_express\"\n                        },\n                        \"zone\": {\n                            \"href\": \"/api/v1/zones/US\"\n                        }\n                    }\n                },\n                \"createdAt\": \"2017-02-14T11:53:41+0100\",\n                \"updatedAt\": \"2017-02-15T13:31:33+0100\",\n                \"_links\": {\n                    \"self\": {\n                        \"href\": \"/api/v1/shipments/21\"\n                    },\n                    \"method\": {\n                        \"href\": \"/api/v1/shipping-methods/dhl_express\"\n                    },\n                    \"order\": {\n                        \"href\": \"/api/v1/orders/21\"\n                    }\n                }\n            }\n        ],\n        \"currencyCode\": \"USD\",\n        \"localeCode\": \"en_US\",\n        \"checkoutState\": \"completed\"\n    }\n"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/orders/{id}/cancel" : {
      "put" : {
        "tags" : [ "Orders" ],
        "summary" : "Cancel an Order",
        "parameters" : [ {
          "name" : "Auth-token",
          "in" : "header",
          "description" : "Authorization bearea token",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "id",
          "in" : "path",
          "description" : "Id of the requested order",
          "required" : true,
          "style" : "simple",
          "explode" : false
        } ],
        "responses" : {
          "200" : {
            "description" : "success"
          },
          "400" : {
            "description" : "already canceled"
          }
        }
      }
    },
    "/orders/{orderId}/payments/{id}/complete" : {
      "put" : {
        "tags" : [ "Orders" ],
        "summary" : "Complete payment",
        "parameters" : [ {
          "name" : "Auth-token",
          "in" : "header",
          "description" : "Authorization bearea token",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "orderId",
          "in" : "path",
          "description" : "Id of the requested order",
          "required" : true,
          "style" : "simple",
          "explode" : false
        }, {
          "name" : "id",
          "in" : "path",
          "description" : "Id of payment to complete",
          "required" : true,
          "style" : "simple",
          "explode" : false
        } ],
        "responses" : {
          "200" : {
            "description" : "Success"
          },
          "400" : {
            "description" : "payment already completed"
          }
        }
      }
    },
    "/orders/{orderId}/shipments/{id}/ship" : {
      "put" : {
        "tags" : [ "Orders" ],
        "summary" : "Ship an Order",
        "parameters" : [ {
          "name" : "Auth-token",
          "in" : "header",
          "description" : "Authorization bearea token",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "orderId",
          "in" : "path",
          "description" : "Id of the requested order",
          "required" : true,
          "style" : "simple",
          "explode" : false
        }, {
          "name" : "id",
          "in" : "path",
          "description" : "Id of the shipped shipment",
          "required" : true,
          "style" : "simple",
          "explode" : false
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "examples" : {
                "tracking" : {
                  "value" : {
                    "tracking" : "The tracking code of the shipment"
                  }
                }
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "success"
          },
          "204" : {
            "description" : "No Content"
          }
        }
      }
    },
...
  },
...
}